+++ spec/factories/app_configs.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 # Read about factories at https://github.com/thoughtbot/factory_girl
 
 FactoryGirl.define do
@@ -18,10 +19,6 @@ FactoryGirl.define do
     overdue_checkin_email_body 'MyText'
     overdue_checkin_email_active true
     terms_of_service 'TOS'
-    favicon_file_name 'favicon.ico'
-    favicon_content_type 'image/vnd.microsoft.icon'
-    favicon_file_size 15
-    favicon_updated_at '2013-06-24 14:02:01'
     deleted_missed_reservation_email_body 'MyText'
     send_notifications_for_deleted_missed_reservations true
     notify_admin_on_create false